==> notebooks/1_0_data_analysis.ipynb <==
# Imports

!pip install torch_geometric
!pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-2.1.0+cpu.html
!pip install torch_sparse

# import required modules
import random
from tqdm.notebook import tqdm
import pandas as pd
import numpy as np
import ctypes
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn import model_selection, metrics, preprocessing
import copy
from torch_geometric.utils import degree

import torch
from torch import nn, optim, Tensor

from torch_sparse import SparseTensor, matmul

from torch_geometric.utils import structured_negative_sampling
from torch_geometric.data import download_url, extract_zip
from torch_geometric.nn.conv.gcn_conv import gcn_norm
from torch_geometric.nn.conv import MessagePassing
from torch_geometric.typing import Adj

# Load Dataset

# load main dataset
data = pd.read_csv("../data/raw/ml-100k/u.data", sep='\t', header=None, names=['user_id', 'movie_id', 'rating', 'timestamp'])
data

# load data about movies
items = pd.read_csv("../data/raw/ml-100k/u.item", sep='|', header=None, names=["movie_id", "movie_title", "release_date", "video_release_date",
              "IMDb_URL", "unknown", "Action", "Adventure", "Animation", "Children's", "Comedy", "Crime", "Documentary", "Drama", "Fantasy",
              "Film-Noir", "Horror", "Musical", "Mystery", "Romance", "Sci-Fi", "Thriller", "War", "Western"], encoding='ISO-8859-1')
items = items.drop(columns=['video_release_date', 'IMDb_URL'])
items

GENRES = ["unknown", "Action", "Adventure", "Animation", "Children's", "Comedy", "Crime", "Documentary", "Drama", "Fantasy",
              "Film-Noir", "Horror", "Musical", "Mystery", "Romance", "Sci-Fi", "Thriller", "War", "Western"]
OCCUPATIONS = ["administrator", "artist", "doctor", "educator", "engineer", "entertainment", "executive", "healthcare", "homemaker", "lawyer",
               "librarian" ,"marketing" ,"none" ,"other" ,"programmer" ,"retired" ,"salesman" ,"scientist" ,"student" ,"technician" ,"writer"]

# load data about users
users = pd.read_csv("../data/raw/ml-100k/u.user", sep='|', header=None, names=["user_id", "age", "gender", "occupation", "zip_code"], encoding='ISO-8859-1')
users

# Data Exploration

data.dtypes

data.describe()

items.dtypes

items.describe()

users.dtypes

users.describe()

# normalization function for user demographic information
def my_hash(x):
  return ctypes.c_size_t(hash(x)).value

def users_encode(users):
  users_enc = users.copy()
  users_enc["age"] = users['age']/100
  users_enc["gender"] = (users['gender']=='M').astype(int)
  users_enc["zip_code"] = users["zip_code"].apply(my_hash)
  max_zip = users_enc["zip_code"].max()
  users_enc["zip_code"] = users_enc["zip_code"]/max_zip
  users_enc["occupation"] = users["occupation"].apply(OCCUPATIONS.index)/len(OCCUPATIONS)
  return users_enc

def encode_users(users, data, items):
  users_enc = users_encode(users)
  return users_enc

users = encode_users(users, data, items)
users.head()

Merge primary data with encoded user demographic data

df = pd.merge(data, users, on='user_id')

df.to_csv('../data/interim/output1.csv', index=False)
